+++ src/main/java/odyssey/backend/domain/auth/exception/UserNotFoundException.java
@@ -1,5 +1,6 @@
 package odyssey.backend.domain.auth.exception;
 
+import odyssey.backend.domain.auth.exception.error.UserExceptionProperty;
 import odyssey.backend.shared.exception.GlobalException;
 
 public class UserNotFoundException extends GlobalException {
